Fixup service worker detection test (#162)
* 🔮 Move service worker event registration to setup.

This enables us to not have to worry about missing events due to late
registratons. This also introduces the concept of a resolved state for
artifacts since they can come in async during the gather phase.

* Wait for service worker to become active.

What we were doing before was taking the first event we got back
which was often in the starting state and not in the active state.