Refactor driver call to about:blank for SW reload. (#250)
Reintroduces `WAIT_FOR_LOAD` (I know) so that we can call a run to `about:blank` and the driver will choose `Page.navigate` or `Page.reload` as it does in other cases. This will mean that the extension shouldn't get into an uncanny valley as it was before.

Sometimes, and intermittently, I've seen the extension hang on the `about:blank` page, because it's been given a `Page.navigate` call rather than a `Page.reload`. So that'd be a reason to use the existing functionality that chooses the correct one to call based on the current URL. But then `about:blank` doesn't fire a page load event ever, so we need to combo that with the `waitForLoad` flag.

Quite what gets the extension into the uncanny valley is a mystery to me, but it definitely happens, and AFAICT this is the resolution.

* Fixes #229