Fix picking CHROME_PATH priority over other matches.
When both regular Chrome and Chrome Canary are installed in their
default paths on MacOS chrome-launcher will pick Chrome Canary due to
the priority weights.

You should be able to force regular Chrome by setting its binary path in
the CHROME_PATH environment variable, but that doesn't work because if
two priorities match the same installation path the first match will be
used, and in the case described above we have two matches for regular
Chrome and one for Chrome Canary. So the first build-in priority is
picked for the regular Chrome and is then compared to the Chrome Canary
priority, and the Canary priority has a higher weight and is picked.

This commit adds the CHROME_PATH priority to the beginning of the list
of priorities, so that when both the built-in priority and the
CHROME_PATH priority match the same installation the CHROME_PATH
priority will match first and be picked.