fix: prevent mutation of default flags (#162)
The default Chrome flags were being mutated by the `flags` method. This
led to changes in the default flags over time, which presumably was not
intended.

The default flags are now shallowly cloned to prevent mutation. The
TypeScript type has been set to `ReadonlyArray` to prevent this from
happening in the future.