feat: Support remote-debugging-pipe and add loadExtension test/example
Historically, the `--load-extension` option in chromeFlags could be used
to launch Chrome with an extension installed. This flag is no longer
supported in official stable Chrome releases since Chrome 137. The
`--remote-debugging-pipe` is the official alternative. To support this,
`chrome-launcher` needs to pass two additional pipes.

This patch also includes a full test that serves as a form of
documentation on how one can load an extension with `chrome-launcher`.

For more context, see https://github.com/mozilla/web-ext/issues/3388

Fixes #289